Prepare existing indexer for use with assigner service
When an existing indexer is configured to work with an assigner service, automatically self-assign the publishers from all of the indexer's providers.

This PR also decouples assignment from allow/block policy, making the two independent of eachother. This makes it easier to reason about what the policy does, whether or not the indexer is using an assigner service. It also means that an indexer's policy does not need to change to work with an assigner service, and allows more capabilities such as being able reject assignment of specific publishers on an indexer.