Allow specification of preset publisher assignments
Each indexer in the pool can have preset publisher ID assignments. This means that when a publisher ID is seen, that publisher is assigned to all indexers that have that ID as a preset assignment. Publishers with preset assignments are assigned to all preset indexers without regard to other existing assignments or to the replication value.

This PR also changes the OnAssignment notification to report the indexer number instead of the indexer's admin URL. Also, the OnAssignment notification will keep reporting assignments of the specified peerID until no more assignments of that peer are needed, after which the channel is closed.