Refined total CIDs indexed estimation
Integrate `/stats` API with the new `indexer.Stats` values when the
backing store supports it. This allows the API to return a more accurate
estimation of the total CIDs indexed.

The logic is cached and refreshed hourly. This is because, such
estimation may take time and can be expensive. Using caching we can
avoid stressing the backing store if `/stats` is called in short
succession.

Relates to #982