Do not auto-assign all publishers
When an indexer is reconfiguerd to use an assigner service, it may already have some indexed content from some publishers. Instead of automatically self-assigning all of these publisher to the indexer, these publishers will instead be seen by the assigner service as preverred assignees for the indexer. This allows the assigner to choose to a subset of the indexers to assign a publisher, preferring tndexer(s) that have already indexed content from that publisher.

When assigning a publisher to an indexer, the candidate indexers are ordered
first by "preferred" status and then minimum number of publishers already
assigned.