Do not auto-assign all publishers (#1091)
* Do not auto-assign all publishers

* Indexers list unassigned publishers, with previously indexed content, and preferred:

When an indexer is reconfigured to use an assigner service, it may already have some indexed content from some publishers. Instead of automatically self-assigning all of these publishers to the indexer, these publishers will instead be seen by the assigner service as preferred assignees for the indexer. This allows the assigner to choose a subset of the indexers to assign a publisher to, preferring indexer(s) that have already indexed content from that publisher. Fixes the previous behavior of simply assigning a publisher to all indexers that had previously indexed content.

* During assignment of publisher to indexer, candidate indexers are ordered :

When assigning a publisher to an indexer, the candidate indexers are ordered first by "preferred" status, and then by the minimum number of publishers already assigned.

* Changed default assignment replication to 1, so that each publisher is assigned to one indexer.
* Additional log messages