Indexer frozen mode and handoff (#1132)
* Indexer frozen mode and handoff

Indexer Freeze

The indexer monitors the remaining storage capacity of the filesystem that the value store is kept on. When the percent of space used reaches the configured threshold, the indexer enters "frozen" mode where it does not store any more index content. It still continues to process ads and handles removal ads. The indexer records the last ad ingested before becoming frozen, and makes this available as part of the providers information.

The only difference when frozen is that the entries are ignored. Ingesting the ad when frozen is necessary to get provider and publisher address updates and metadata updates for existing index content.

Indexer Handles Handoff when Assigner Tells it to

The assigner service tells an indexer to take over ingesting ads from a publisher that was assigned to a frozen indexer. The indexer handles receives the publisher being handed off and the ID and address of the frozen indexer previously handling the publisher. The indexer then queries the frozen indexer for provider records, finds the provider that has the handed-off publisher, sets the last processed ad to the ad that the provider was frozen at on the frozen indexer, and syncs with handed-off publisher. The sync only goes as deep as the last advertisement the frozen indexer processed before entering frozen mode.

Assigner Service Detects Frozen Indexer and Initiates Handoff

The Assigner polls indexers to detect when they have become frozen. On finding a frozen indexer the Assigner selects a candidate indexer for each assigned publisher. The Assigner then tells the candidate indexer to do handoff giving it the publisher and the frozen indexer info. If handoff is successful, the Assigner changes the assignment from the frozen indexer to the new indexer that the publisher was handed off to, and then the Assigner does the same for the next publisher to handoff. If handoff is incomplete, the Assigner will attempt to handoff the remaining publishers at the next poll.

 Add unfreeze functionality

The indexer can be configured to revert back to unfrozen mode at startup. This reverts the latest processed advertisements to what they where when the indexer was frozen, only preserving the any updates to the publisher and provider addresses.

Additional items:
- Log disk usage and warn before freeze
- ingest.Sync does not start its own goroutine and return error
- Additional tests of freeze/unfreeze