Add disk usage metric for value store (#1139)
* Add disk usage metric for value store

Maintain a disk usage metric, percent usage, for the file system located at the location of the value store.

This addresses #119 and #484

* test speedup

* Remove useless metrics update signal

Ingestion metrics are updated periodically if an update has been signaled. If the indexer has any activity, update is signaled. So in almost all cases the update is signaled. If the indexer has no activity, there is no point in avoiding the update because the indexer has nothing better to do and then the metrics should still be updated as they may be affected by other external factors like fs usage or storage compacting.

So, remove the update signaling since it does not help anything, and only causes unnecessary context switches. Simply update the metrics periodically.