Save advertisements as CAR files
When configured to save advertisements, save advertisements and entries from the datastore as CAR files in the configured directory. The advertisement and its entries are removed from the datastore after they are saved as a CAR file.

If there are existing advertisements saved in the datastore, on startup, save these as CAR files and remove them from the datastore.

Advertisements are not saved when the indexer is frozen.