Save advertisements as CAR files (#1210)
* Save advertisements as CAR files

When configured to save advertisements, save advertisements and entries from the datastore as CAR files. The advertisement and its entries are removed from the datastore after they are saved as a CAR file. If there are existing advertisements saved in the datastore on startup, then save these as CAR files and remove them from the datastore.

Advertisements are not saved when the indexer is frozen.

The CarWriter uses a `filestore.Interface` to seamlessly support different file store implementations (local filesystem, S3, etc.).Current implementations support saving CAR files to the local filesystem and to AWS S3.

Advertisements and entries are stored in separate CAR files. This allows downloading advertisements withing having to download the multihash entries, and is useful when loading advertisements whose entries have already been deleted.

- If the advertisement CAR already exists at the destination, do not re-write it.
- If the entries CAR if it already exists at the destination, do not re-write it.
- Allow various S3 parameters, normally set by the environment to be overridden:
  - Region
  - S3 Endpoint
  - AccessKey, SecretKey

Unit tests use the same code to test the different filestore implementations, demonstrating that they behave the same and their use is interchangeable.