Deploy translation of all non-streaming lookups to streaming on `dev`
Deploy the latest `indexstar` that introduces translation of all
non-streaming lookups into streaming backend requests.

Increase backend call timeouts so that the chances of at least cache
getting updated even if client times out earlier.

In the case of lassie the timeout is set to 1 minute, and short timeout
on server side results in false-negative 404s.

Once tested on `dev` will be rolling out to `prod`.