Fix a bug where unknown paths always return landing page
The find server would fallback on landing page on any request for which
there was no matching path, because the handler for `/` was not checking
path. This meant that requests for paths that are not offered by finder
like `/metadata` would result in 200 response of content type text/html.

Instead, find server must check the path requested on fallback mux and
only return landing page if either `/` or `/index.html` is requested.

The changes here strictly check the path before returning the landing
page and assert it is so in tests.