Write advertisements and publisher heads from datastore to carstore
At indexer startup, the latest processed advertisement CID for each publisher is writen to the car store as a head file. Then all existing advertisements are read from the datastore, and in the background these are written to the car store as CAR files.

During normal operations, when the latest advertisement is written to a CAR file, a head file is written for the publisher as well. This head file is later used by other indexers when importing CAR files, as a place to sto importing directly from the publisher to switch to importing  advertisements from the carstore.