Move `caskadht` to CPU optimised worker nodes
Instantiate two other worker nodes with cpu optimised instance types to
have one per AZ in addition to the pre-existing one.

Move caskadht to run on those worker nodes with as much CPU and RAM we
can give it.

This should help identify the limits at which the accelerated DHT starts
behaving until fixes deeper in libp2p stack are made.

Relates to:
 - https://github.com/ipni/caskadht/issues/28