Remove code for registering providers by SP ID (#1336)
Discovering/registering index providers, using an SP ID, is a feature that was never used. It created unused code and API. This functionality and its "Discovery" API has been removed. This functionality was removed from the registry and server endpoints.

The libp2p ingest server was also never used, and it too has been removed. The ingest server is currently only supplying the /announce API for direct HTTP announcements, and these are not sent over libp2p

Looking up a service provider's p2p ID and address may still be useful when trying to match them to a provider reported by the /providers API, or when determining the address to perform a sync with. For this reason, there is a new command "spaddr" that retrieves the SP addr info when given the SP ID. This reuses the lotus client functionality that was part of the removed Discovery API.