Ingest advertisement data from car mirror
When ingesting advertisements, get their multihash entry data from car files in the mirror if available instead of syncing the entry chain with the publisher.

Since the vast majority of advertisement data consists of the multihash entries associate with each advertisement, the functionality demonstrates a significant reduction in data transfer between the publisher and indexer.

- Changed configuration to configure AdvertisementMirror
- An advertisement mirror can be configured to be readable, writable, or both.
- Changed how CarReader reads entries to fix possible goroutine leak
- Fixed check for existing files in CarWriter.
- Do not write head file if CAR file already exists.
- Unit test to check that ingesting multihashes from mirror works