HTTP sender adds p2p ID to addresses in announce message
This is necessary to communicate the p2p ID when using HTTP. Previously this way done outside the HTTP sender, and needed to be done for every announce message sent using the http sender. This needed to be done by the sender. So, now the http sender is given the peerID when the sender is created, and this is added to the addresses in each announce message.