Add get pending syncs command
Currently there is no functionality for checking whether there are any syncs in progress. That complicates troubleshooting, specifically when the idexer seems to be busy with doing something without writing anything into the datastore. This commit ads a new admin endpoint and respective CLI functionality for fetching a sorted list of currently pending syncs. It also prevents a sync for the same peer from being submitted while there is one already in-progress.