Publish ECR tag on `version.json` change
This is to work around the recent changes in the unified CI, where it
also creates the github release when the version file is updated.

As a result, because release is published by a bot the ECR image build
with the release tag is not triggered.

The changes here would trigger image publish as soon as the version file
is changed, not relying on specific github release events.