Add `gp3` storage class with 5K IOPS
The expropriations on `dido` volume showed that 5K IOPS with 300 MiB/s
throughput on `gp3` volume type providers sufficient performance.

Define a storage class with the required IOPS and throughput in order to
move production nodes to it.

Instantiate fresh snapshots for nodes as the latest one is 4 days old.