Fix OOM by not keeping all ads in chain in memory
Long advertisement chains are iterated and every advertisement is stored in memory until after the chain is processed. This is causing an out-of-memory condition and leading to crash/restart of indexers. Fix this by not keeping all advertisements in memory at the same time and instead loading each advertisement at the time it is processed.

Fixes issue #1569