Fix deadlock with manual sync (#1580)
A manual sync creates a channel that adProcessed events are written to. If multiple events are written to that channel before it is read, this can cause workers to block waiting to write to that channel. If workers are blocked, they will not read SyncFinished events. If SyncFinished events are not read, then dagsync syncs will not complete. When this happens the original manual sync may not finish and the channel will not be read, thereby entering a deadlocked state.

Fixed this by having manual syncs create a channel that cannot fill up and block. This will also help prevent blocking other workers that trying to write adProcessed events.