Handle trailing slash with find batch request
The findBatch handler is called when there is no slash following the /multihash endpoint. If there is a slash, then the findMultihash handler is called and expects a multihash in the URL path. Some clients may add a trailing slash, even when they mean to call the findBatch handler. Fix this by calling findBatch if findMultihash is called and the request method is POST.