Deploy deletehack to dev
deletehack is a dedicated storetheindex instance that is going to go through the ads from s3 mirror and apply only deletions to dhstore. This is required to compensate for deletions that didn't use to be applied in the past. After deletehack has caught up with the rest of the cluster the instance will be deleted.