Use latest address available in ad chain (#1669)
If the first advertisement does not have a provider address, then use the next on available in the ad chain.

This fixes a problem when ingesting long ad chains where the first advertisement received is missing the provider address. When this happened, the indexer would revert to its old behavior of updating the current provider address to whatever is in the ad being ingested. This may be an old and no longer valid address that would not be updated to the current address until ingesting ads later in the chain. This fix uses the latest provider address in the chain, even if the first ad is missing the provider address.