Return Not Found message with 404
When an incorrect URL is given to the find API, a blank response body is given with the 404 error code. When used with a user interface like a browser or curl, this is confusing until the error code in the response is inspected. Returning the "Not Found" message eliminates the confusion.