Explicitly set dhstore stateless resources in `dev`
In order to have a deterministic and isolated performance analysis
explicitly set the resources for `dhstore-stateless` instances. This
would make it easier to reason about potential bottlenecks or their
absence in the ingest pipeline, since pods would be given the same
resources on every deployment and cannot outgrow the limit specified.