Increase max outstanding IO operations to 256
FoundationDB used AIO with O_Direct interface to interact with disk. The
maximum number of outstanding IO operations default to 64. If this value
is too low then the ratekeeper limits the accepted write rate which
results in a growing lag between memory and disk in log and storage
servers.

Bump the max outstanding kob to 256 in order to observe the effect on
lag.