Disable index counter on `inga` to restore service in `prod`
It seems locking contention in leveldb used by index counter causes high
CPU spikes that then results in readiness check failure and ultimately
cause all lookups to cid.contact to fail.

To mitigate the issue and restore service, deploy a quick fix where
index counter is completely excluded from execution path until
appropriate fixes are in place. This will assure that the service is
restored in the meantime. However, it means that we may end up with
inaccurate index counts for some providers.

This is an acceptable tradeoff since absolute index counts are not as
important than lookup service itself or their relative growth.