Recover FDB prod cluster from too many old generations
Self-recovery of the FDB nodes on prod was significantly delayed such
that it halted ingestion. This was caused by growing write lag on the
nodes and eventually knocked out the coordinators which resulted in the
cluster being unresponsive.

To fix this, specific storage nodes with large lag were excluded, and
the max generations was changed from 0 to 200.

The changes also introduce dedicated ratekeeper nodes to reduce load on
storage servers.