Reduce compaction thresholds back on `qui`
After experimentation, it seems parallelism of 20 on the current node type that qiu is running on yields fast enough compaction debt catch up at the rate of ~40GiB per hour.

Reducing compaction thresholds back down to keep compaction as aggressive as before. We may need to reduce these further depending on write load. For now the limits seem to result in a stable read amplification (on helga) which uses the same thresholds as default values.