Shut down FDB and indexer backed by it on prod
After migration to the latest FDB version the data migration is still
ongoing. This has taken about 2 weeks and it is unclear how long it will
take for the backing store to be upgraded to redwood.

Due to the lack of engineering bandwidth and the amount of time it takes
to investigate the issues further shut down the foundation db and
revisit at a later date.

Change the assigner to use a fake indexer to avoid crash loopback due to
the strict requirement that at least one indexer should be present.

See:
 - https://github.com/ipni/storetheindex/issues/2168