Replace registry goroutine and channel with mutex
This simplifies the registry code and makes it more efficient.

Originally, the registry was going to run many functions asynchronously in its own goroutine so that callers would not have to wait for the processing to finish. Over time, all the calls to registry functions became synchronous, so the channel and goroutine were nothing more than a synchronization mechanism. This PR replaces that goroutine with a simpler mutex.