Increase http timeout to 3m
When a request for ad entries if made, then advertisement provider builds its entries chain at the time the request is made. If the entries chain is long, as it the case for at least one provider, then the response is not sent for some time. This may cause the indexer to time out waiting for a response.