Explicit sync continues from last fully-processed ad
An explicit (manual) sync will do nothing if all the advertisements have been synced, but their entries have not yet been synced. This means that either a new advertisement or an explicit resync is required to process the partially synced advertisements.

A common reason for doing a manual sync is to restart a sync that failed, or was interrupted due some error. A new advertisement may not be available to restart processing the advertisement chain, and ipnisync will not send a new advertisement notification because it has already seen the latest advertisement.

Auto-sync suffers from the same problem. If a sync is interrupted after the advertisement chain is synced, but before all the entries are synced, then subsequent polls will do nothing until a new advertisement is published.