Add IPNI garbage collection functionality
A new ipnigc command walks a provider's advertisement chain to find advertisements that have a deleted contextID. GC then deletes the indexes from thost advertisements as well as the correcponding CAR files from the mirror.  Empty and removal CAR files are also deleted from the mirror. This curronly only works with a DHStore deployment as is runs stand-alone from indexers.