Identify ipni-gc resources by provider ID
Previously the publisher ID was used as this was more immediately associated with an advertisemetn chain, and because head files in the CAR mirror were identified by publisher ID. It is more correct to change these things to be identified by provider ID, since it is possible for the same cprovider's hain to be served by different publishers.

This PR also allows non-existent provider resources to be cleaned up even if GC has been previously done.