Process ad chain in segments during GC. (#2460)
* Provess ad chain in segments during GC.

If the amount of the ad chain to process is very long, then do not process all the ads at the same time. This may lead to keeping too many CIDs and context IDs in memory resulting in an out-of-memory condition. Instead process a segment of the chain at a time, starting from the oldest to the newest segment.

* Remove commit flag
* Use default temp dir if not configured