Poll providers that have sync errors
A number of providers do not complete processing advertisements due to various errors, such as network timeouts, rate limiting, etc. Processing can often continue when retried after a short amount of time. However, without a announcement of a new advertisement, the provider will not be retried until the polling time has elapsed. This change includes providers, with errors during the previous sync, in the set of providers to poll.