Remove untick in exchange for increments larger than one
Remove untick from vector clock at the risk of having increments larger than one for the local vector clock in edge cases where persistence of snapshot may fail in the midst of a cycle.

This is an acceptable risk considering the difficulty in implementing transactional IPLD node storage and the application of it in this particular case.