Configurable option to skip ads with 500 error on first entries block
A new boolean config option `Ingest.Skip500EntriesError` allows the indexer to skip advertisements when the publisher returns a 500 response with "failed to sync first entry". This value is run-time reloadable.

The purpose of this is to skip ads that to publisher will never deliver content for. There is an issue with providers where they will not return entries from some advertisements, and permanently return a 500 result when trying to sync the entries.