Expose pebble config option PebbleFormatMajorVersion
The config value `Indexer.PebbleFormatMajorVersion` sets the format of on-disk files. It is recommended to set the format major version to an explicit version, as the default may change over time. When unset or 0, use the current format version. When set to -1 update the format to the latest supported  version.