Ignoring configuration for nonexistent logger
The libp2p logger "basichost" no longer exists, so trying to set a log level for it causes an error. This error is properly logged and skipped by storetheindex, but causes the assigner to fail. This PR fixes the assigner to correctly handle trying to set the log level of a logger that does not exist.

Additional documentation updates are included in the PR.